# docs/swagger.yaml
basePath: /api
definitions:
  entity.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      text:
        type: string
      user_id:
        type: integer
    type: object
  entity.CommentCreate:
    properties:
      post_id:
        type: integer
      text:
        type: string
    required:
    - post_id
    - text
    type: object
  entity.CommentUpdate:
    properties:
      text:
        type: string
    type: object
  entity.Post:
    properties:
      created_at:
        type: string
      id:
        type: integer
      task_id:
        type: integer
      text:
        type: string
      user_id:
        type: integer
    type: object
  entity.PostCreate:
    properties:
      task_id:
        type: integer
      text:
        type: string
    required:
    - task_id
    type: object
  entity.PostUpdate:
    properties:
      text:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.refreshTokens:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.signInOutput:
    properties:
      access:
        type: string
      refresh:
        type: string
    required:
    - access
    - refresh
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for TodoApp
  title: ToDoListApi
  version: "1.0"
paths:
  /api/comments:
    post:
      consumes:
      - application/json
      description: Create Comment
      operationId: create-comment
      parameters:
      - description: comment body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.CommentCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Comment
      tags:
      - comments
  /api/comments/:id:
    delete:
      consumes:
      - application/json
      description: Delete Comment
      operationId: delete-comment
      parameters:
      - description: comment id
        in: path
        name: input
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Comment
      tags:
      - comments
    patch:
      consumes:
      - application/json
      description: Update Comment Info
      operationId: update-comment-info
      parameters:
      - description: comment id
        in: path
        name: input
        required: true
        type: integer
      - description: updated comment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.CommentUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Comment Info
      tags:
      - comments
  /api/posts:
    get:
      consumes:
      - application/json
      description: Get All Posts
      operationId: get-all-posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Post'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get All Posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create Post
      operationId: create-posts
      parameters:
      - description: post body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.PostCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Post
      tags:
      - posts
  /api/posts/:id:
    delete:
      consumes:
      - application/json
      description: Delete Post
      operationId: delete-post
      parameters:
      - description: post id
        in: path
        name: input
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Post
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Update Post Info
      operationId: update-post-info
      parameters:
      - description: post id
        in: path
        name: input
        required: true
        type: integer
      - description: updated post info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.PostUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Post Info
      tags:
      - posts
  /api/posts/:id/comments:
    get:
      consumes:
      - application/json
      description: Get Comments of Post
      operationId: get-comments-post
      parameters:
      - description: post id
        in: path
        name: input
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Comments of Post
      tags:
      - posts
  /api/refresh:
    post:
      consumes:
      - application/json
      description: Refresh tokens
      operationId: refresh-tokens
      parameters:
      - description: refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.refreshTokens'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.signInOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Refresh
      tags:
      - auth
...
  /api/token:
    post:
      consumes:
      - application/json
      description: Login in account
      operationId: login-account
      parameters:
      - description: account login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.signInOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Login
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"